--- typst-package/example.typ
@@ -1,7 +1,0 @@
-#import "./lib.typ" as tiaoma
-// #import "@preview/tiaoma:0.1.0"
-#set page(width: auto, height: auto)
-
-= tiáo mǎ
-
-#tiaoma.ean("1234567890128")
